pyproject.toml
[project]
name = "mlops-mlfactory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.69",
    "fastapi>=0.135.1",
    "mlflow>=3.10.1",
    "pandas>=2.3.3",
    "prefect>=3.6.22",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
]
